created the apis folder and created a tester application setup for how to connect my api to a front end.
The next steps are figuring out authentication, and setting up my databases, as well as then connecting the database to the other code.

I also need to figure out how to move all of this code into a docker container, so that I can get the cloudflared stuff working, so that i can have the api serving over https, on a domain name.