TOPLCD creation
Have now gotten the very basics of the top LCD in place, as well as working on making the top area more responsive, so that when the window gets squeezed, it now has the bar filling that space, rather than some rather squeezed volume and playback controls

next up will be prettifying the lcd, geting the images working, as well as adding in the other functionality, such as the playback scroll through, and all the different text. And one of the more exciting aspects will be seeing how well i can get the menu system to adapt here, because in theory, it should go without a hitch?