UPDATE chipmunk
I have now got all of the image compression stuff working, and when i go and access the images i can see that it has been the right decision since the images can back down so much quicker than otherwise, all the way upto the 640*640 image it loads instantly.

So next Up I will either implement uploading the details to the database,

or the systems that will allow me to delete users and their files at will, which i spose is more of a dev thing, but it really would be annoying to have to go through and manually delete them, so i want it so that when i delete a user row, that will trigger a system to go through and delete all of their other entries in the database, and then to get to the files table and check if a file has multiple owners or not, and if there are files where they are the sole owners, it will delete those files off of the drive, and I obviously don't want to get that one wrong so will be careful with that.