UPDATE admin_tools
I have taken a little detour with the admin tools to add a table of users before implementing the delete user function, since this table  will be needed anyway and I might as well make a rudimentary version here and now before I update it later on, since it will be useful here and now for viewing the users that i will be deleting for testing, and at some point i can make a proper tool that will allow me to make more of these tables in an easier way, so that i can implement one for the files as well as.

It should also be said, that this isn't my final intended design for the admin page, and or the way in which users are deleted, but it's just as stated earlier a stepping stone for getting all of the other development done, and then admin tools.

And i think by making a solid admin tool, i can avoid shipping pgadmin in prod, and ship this easier interface instead