UPDATE admin & signup
signup: I have changed it so that there is now a spinner and confirmation/error indicators, however tomorrow I will implement the visible error messages, i might just use the browser alert system, because i'm lazy. It also now redirects the user to well currently a different port on localhost, however when deploying it it will just go to the root of the domain

admin: I have updated it so that when you delete a user,  it clears the field, so that you don't have to do it manually, and at some point i will just make a system whereby you can just specify an input, and then it will automatically add the remove indicator, and add clear paths and such. In the future it will be...