CREATE projectView
I have now started working on making the project view. Currently have a function which will load in the html to  where it should be for the project view, and i will now start working on making the actual styling for this. I think I will refrain from having a seperate project start page/form, and will instead make the different fields just built in as natural elements, like being able to double click the title, and changing the value. This will naturally require more work so that it updates in the table properly, but it's the best way of doing it i think?