UPDATE login/auth
Have now fixed a bug where because in prelogin only one value was being returned from  verify_jwt, it couldn't return a false when the JWT was no longer valid, and so now that has been fixed, meaning that it will now propmt the user to login, when the jwt has been verified as false.

At some point i should update the system so that it will catch authentication problems, and then prompt the user to relogin, but that will come later.