UPDATE project view
Have now implemented the x close to be a button with img svg, but i don't have the right svg at the moment, which is unfortunate.
Also changed it so that the title and artist name and year and such were coming in as regex replacements, instead of hard coded.
And have started working on the system whereby the user can edit the description using markdown