UPDATE project view
Hooked it up to the menu system, and now the menu atleast displays, so next time will be making the table that is a list of songs, and implementing good systems for adding in new song, and whatever, since we have never designed that before. Also just being mindful of what data is actually needed, since the project does not need to have the name of the project repeated throughout, since we obviously know that it is this project. Many interesting things to keep a head of