UPDATE project view
I have now gotten the drag and dropping of files in working, and next up will be figuring out how best i can go and actually transfer these files to the server, and buildingn the json files which will then live in the database, and file server. And then i get to go and do the fun stuff, which is the processing of audio files, compression and that sort of thing.