UPDATE audio upload
I have now gotten the different stages of uploading the data to the database working, i still need to assign the async split and stuff, but for today i am happy, and i have just listened to 'when i kissed the teacher' being sent over the wire from nginx, to me, by accessing it at om2media, and I must say i really am happy that this is working. I have much work ahead of me, but the very core of what i need is almost there, and i think with a little bit of work i can have projects working very soon. Atleast the system of serving up the audio is much better than it ever was in om1, i can select the quality of the stream at will now, and i guess that will be useful for users with limited broadband network, or when they are doing it over mobile...