UPDATE projectView / routing
I changed the order in which new project ID's should be generated, since it wouldn't do for /projects/ to be a real route that created a new project every time, because then when people navigated back it would create a new project, and that would be bad, so instead i have decided to capture the route /new-project/ and that will then not be turned into a route, and instead run into the function that will initialise the project in the database, and that will then return an ID, and then I can set the route to /projects/id... and that will then trigger a real route, and that will also work perfectly with forwards/backwards navigation.

But to achieve this I had to change up the handleRoute function so that it would actually worked with dynamic routes, and it does, which is great.