UPDATE playback.js
So, With this commit we now have audio playback working when you press the play button inside of a project (the button with the text Play inside of it). This is great, because While at the moment it will only play the first track, and it doesn't display any of these details, and yada yada, we've gotten this far, and it won't be astrononmically much harder to get consecutive playback working, and well that's what i am going to get on to make now.