UPDATE playback / playback controls
So, I have now gotten the play/pause button working, but without the error handling for when there is no song in the queue, and I have gotten the next button working so that it is possible to go forward in the queue. This is really good, because now you can play music properly (the other things are of course wanted, but these core things are now working and existant). And of course when one song finishes, you are now moved forward to the next song. I guess my rush to get this part of it done has perhaps resulted in a little bit of spaghetti code, but this can easily be cleaned up later, thanks to my modular design, as opposed to the last one which was created in the most terrible fashion. Also I want to implement the function that will update the top display now.