UPDATE musicObjectsGrid / main.py
I have now gotten the deletion of projects working, there are a number of things that will need to be changed later, for example implementing the deletion of the image when the user is able to upload an image to their projects, and at some point in time i might create a sort of feature where you will have recently deleted stuff, but that will be in the future, since it isn't a core functionality, and instead would be a nice to have. I think the only thing i would have to do to implement a delete later thing would be add a value to their database row, which could be delete/hidden/visisble, and then also a deletion date, and so then 30 days later it would go through the actual deletion sequence, but the user should actualy be able to order the deletion before, but that's all for later.