UPDATE music objects grid
I quickly implemented the play button on the music object grid, and well it was basically a case of copying over the code that triggered it from projectView, and it worked, which is really very good. The main things to implement with playback now is shuffle, adding to the queue, and getting queue interactions working, because otherwise we are nearly there. (also getting top to hide itself when there is nothing to display)