UPDATE deployment
I have been working on getting the project close to being deployable, and in this commit that means that I have been working on creating an init script for the database, checking if the database exists, and it's tables, and if not creating them. I have also modified the startup function in main.py so that it now checks if there exists an admin user, or a default user, and if not creates it from the creds specified in the environment variables.