UPDATE fastapi
Have now implemented it so that in the startup function, it will check if the admin password which is set in the .env matches with the one in the database, and if it doesn't it will be updated to be the correct one. This means that you won't be locked out if you change the admin password in the environment variables file, after the service has been initialised, plus it means that there is now an update password function, which can be used when that functionality comes user side.