UPDATE mobile styling
So I have now done more mobile stylings, but am still in the process of it. In this commit I have moved the nav bar down to the bottom, and have moved two of the buttons over to the right, instead of them being to the left, and then hidden the other three, the'll be accessible from somewhere else. And next it will be styling the LCD in a way which makes sense on mobile, and then I will also have to add some new html for mobile only, because there are some things which make most sense on mobile, and I will have to add back the home buttons that I have removed in this process.