UPDATE mobile styling
Have now styled the bottom bar as far as it will be styled today, and that means that there is an image with rounded corners in the bottom left hand corner, and the title text is there next to it, and then over on the right hand side are the play/pause button and the next button. And now I will add the sticky bar for projects.