UPDATE music object grid
I have now added long press on music objects grid detection, and this means that on mobile you will now be able to long press, and that will mean that you can get the menu up. And I have also added a vibration for this evenet, so now I will see if that is something that interests me going forward.

Also I built it so that i can now deploy it and see how the effects really are.