UPDATE project view
I have now updated project view so that when the full description/notes is shown it will only change the zindex when on mobile using a media query. As well as that also set it to return to 1 instead of 0, and I hope that will fix an issue where the menu on mobile whould show behind the nav bar, rather than infront of it. And lastly I have also increased the build version to 0.0.3, because I am quite partial to the last one, since it finally got mobile good.