UPDATE menu / project view
In this commit which is perhaps a little bit to big, I have modified the menu system so that it can now support more different menus, and at the moment the only one is the one that I am adding now, which is the image selector/updator for projects.
And I have now styled the image selector/updator, and now I have to add the reactivity for the menu, so that users can actually upload an image, rather than stare at a non functional menu.