UPDATE music object grid
I have now gotten the image url properly loading on the main page, however the ones that haven't had images selected don't look so great, so Soon I will have to create a system that allows for a dummmy image, and then building that into every section that uses images. I think the easiest path would be to add an image which would have an url and then the extensions such as /1/2/3/4/5, however that brings it's own complications, so that's why it will be done after we have gotten the other things all ready.