UPDATE menu / project view
I have now gotten the updating of the backgroud images on a successful image upload working, I had done a very lazy for updating it first, but I actually went through and did it properly, and well it's obviously a better way of doing, so now I actually target the top image and change the src, whereas before I just hid and re showed the view, which is bad.

Now I need to make sure that the styling all works on mobile, as well as adding a loader of some sort.