UPDATE menu
So I have now added a loader to the image select menu. And I have added the  mobile styling, which was pretty easy to implement, just adding two media queries, and basically just making it bigger, and moving it down, and hey presto. Further more the menu style that I have created for this specific menu is quite good, so I think I will use it for the other menus now.