UPDATE queue
Now I have gotten the queue updating when the user drags and drops an item in the queue. And I had been dreading this, but actually it was really easy, to implement this using the new queue, rather than the old array, I just had to swap to the new array. And that was all.