UPDATE playback controls
I have now added it so that the play back volume will carry over even if you refresh the page, because now it is stored in local storage. The next thing in this sort of thing would be total sync, however that is not for now, there are more important things to finish first.