CREATE playback sync
I have now added the playback sync functionality, so that when you open om2 again on a different device or something it will have the same state as when you left it on the other device, and i think it is really cool. And now I need to try how it is to use on different devices, and wether it is any good at all.

I also added a few new helper functions, and i will be removing a few of the functions in main.py, because they do almost the same thing, and as such I have function now whch will cut down on that repition.