UPDATE playback/lcd mobile
I have updated the styling of lcd/mobile, so that the backdrop filters should display properly, and so that the text's won't overflow. And I have updated playback so that there is now a playing state, and if it is false, then when playback start happens, it will check if the audio has been playing, and if not won't start the playback, only if it has been given the directive of starting playback.