UPDATE styling
I have now created a sort of stop gap solution for the top bar styling when there is limited space, so that it still works. And I have gone ahead and added -webkit- to all of my backdrop filters as well. So now It should all be good for that aspect on safari, however the weird issues with the square outlines which should be round is confusing, perhaps there is some way of mass fixing that?