UPDATE . / CREATE routing.js / update sync
Tidied up a few things, moving the routing and syncing code out of main, and into it's own module, and have updated the syncing so that it happens in said new module, and so that it keeps a local hash of the last state update, so that if the current one matches, it won't send up an update to the cloud. Which means means fewer excess/dumb network requests, which is allround good.