UPDATE readme, deploy.md
I am now working on upating the readme, to make it better, and right now I need to see if the link I added will properly work, because it is relative.

And I changed that order that building/and deploying was ordered in deploy.md, since people using the thing, but not building it won't care about the build step.