UPDATE the configurable URL's and stuff
And I have now gotten it so that you can change the path that things are being served at by an environment variable, and well it works, and all I had to do was read somebodies stack overflow answer. That was pretty easy. And it works deployed too. Although admittedly I just updated the 5 or so lines that were affected over there, and then if anybody uses this they can do it through there now.