UPDATE styling
I have now gone through and updated the menus and the queue now, and it really does look quite good I think, especially for the queue that was looking a little bit weak before, and now it looks much better. There might be issues in some places with enough contrast, however I will look into this as I progress with this still update.