UPDATE styling - nav bar
I have now updated the top bar so that it is no longer transparenent, and it is now a solid colour, and I have chosen the just background colour, and now the sort of outline colour for the bar is the accent colour. And it looks quite good, and I also updated all of the filters, so that they now use the correct colours.