UPDATE styling
I have now updated the MORE description button so that instead of using that incredibly inefficient system of hiding characters, it now uses  a before that has the background colour, and it filter blurs that, and then the javascript that was involved has been removed. So basically a win win, however I can no longer specify a background colour for the more button, but that's a small thing.