UPDATE projectview/upload
I have now updated the upload to no longer do all uploads at once, and rather do them one at a time, this should really have been addressed earlier, but it was a smaller thing.
However it is possible that with a lot of files it will eat a lot of memory with those recursions, but I'll see how that goes with time.