UPDATE loadAccountImage + UPDATE menu
I noticed that it was using it's own function for getting the users image, and that simply won't do, so I just updated the network_requests function getUserDetail to also get user images, and well that's not a hard thing, but it means that there is now less code doing network stuff.

And it is now possible to sign out.