DELETE audioPlayer.js / UPDATE  om2.js
I removed audioPlayer.js since it was the old one, and all the code that was needed from it has been moved over and improved, and it's now obsolete.

Upated om2.js so that it will now have a formatTimeSeconds function, which takes seconds as an integer and returns a string of formatted time. And I will also add in the format days function which will take an integer in and output that in a Now/days/week/months/years format.