REFACOTOR / UPDATE routing
I have now added a 404 handler into routing, so that if the user navigates to a page which isn't available it will go to the 404 page, however it doens't work for project pages yet, so that will be fixed soon.

Plus I have gone through a lot of code and refactored it so that it is in a good format and doesn't contain crud.