UPDATE settings admin
So in this commit I have now gotten the table in and it is properly done up, and gotten it so that there is a checkbox where you can select the row/(s), and it will then add those selected to an array of uuids, and this means that when I then add the buttons for delete user, and promote to admin, and demote from admin and such it will be easier to have this interface, and perhaps also if we make it so that when only one user is selected it will have their display, I'm not quite sure but that might be the done thing.

Anyway what this will mean is that I have removed the functionality to double click the uuid column for that value, since it seems a little bit silly/jank when I can have them stored in an array instead, so this will be the way forward I think.