UPDATE apis/web
So This morning I have made it so that everything is now served from go/gin rather than through 2 seperate nginx containers as well, and I have made it so that only one domain is used, and that will now just be like om2.domain.tld, instead of the seperate ones for om2apis and om2media, because that's hardly going to be the most meaningful thing for me and the other 3 people that will ever run this. But now that I have that house cleaning out of the way it will be time to get back to writing out the api paths again.