UPDATE apis/web
So This morning I implemented HTTPonly cookies, which also means I need to find some other alternative for how to access/use the local development server version of the web end.

But It does now use cookies, which means that I don't have to use post requests for the things where previously I was using post requests, and that is a great thing.

Plus I am half way through implementing get_project_details, which will allow the project pages to be loaded.

However I did turn off the geo checking of JWT's, since it was making what otherwise took 5ms, take 500ms, and that is not okay for every single request.