UPDATE apis/web
So I have now gotten it so that I can develop locally with cookies in a way which isn't entirely frustrating, it means using the vite proxy, and not using http only cookies in dev mode, but that is fine, and then in prod mode I will implement a switch so that it does use the HTTPonly cookies at that point.

And I have now gotten that most basic of functionality implemented now which allows the loading of projects and play back from them. And soon I will get the other missing routes back and implemented, and that will mean that I can go forth and change the things that I wanted changed, such as changing the structure of how a project json is stored, moving it over into it's own table, and then making it so that the structure is just compiled back to how it wants it to be from the frontend.
This will allow for the editing of project once it has been created, meaning new Song titles can be used instead, plus I will be able to add the features that I really want, like having multiple versions of one song in a project in an orderly fashion, but everything in its time I guess.