UPDATE apis/web
So I have now gotten the creation of new projects working.

And I have gotten the loading/getting of static assets which aren't directly linked to the front end working. And the fact that they are baked into the go back end is so much better than the way that it was done before were they were manually loaded into a pre made nginx container, through the docker compose, which was a little bit silly, that is now rectified, which means it should be much easier now for users to start up om2. There are still things that I need to work on towards that regard, however we are getting there. The smorgasbord of neccessary small files to get the older version of om2 running was silly. Now I want it to just be a docker-compose and an env file.