UPDATE menu.js + go
Firstly I fixed a bug that I hadn't previously known about where the image picker would upload multiple images, if the user had gone through and selected multiple images, however that is now fixed and it will know only pass in one image to be uploaded to the image uploader.

And I also got the unique url string generator in place, so now it is time to start finding out how I manipulate images in go.