UPDATE apis/web
I have now finally after quite a bit of work gotten the full uploading of images working, and it's for some reason just been quite frustating getting all of the small details working. Especially my dumb decision to use json[], when I could have used literally anything else... json[] is a nightmare to get working, but to be fair it does give some very concise queries at times, which I am sure would have worked just as well in text[].

The next step is to get audio upload working, and to add a submit button to that, and then it's just the two signup apis which I will be merging into one and I will have full feature parity with python (oh and also that admin table route, but it is a simple get), and then it will be time to go further than before.