UPDATE apis/web
I first of all added a proper submit button to the files selection area, as well as making that area properly keyboard navigable.

And then I have gotten almost the entirety of the audio processing pipeline sorted, the only last thing that I need to get into place is modifying the project JSON to get the data in there, or perhaps now is the perfect time to just go and implement the changes that I was going to make. I think so.