UPDATE apis/web CREATE marketing/website
So I decide I should probably implement an authorisiation step after the authentication step. And I'm about half way through, but it's so boring, so I also decided to start working on the marketing website. And that's a lot more fun. I should probably not do too much more at the moment, but there's a good shell there, and something to build upon in a few weeks when everything is all nice and done.