UPDATE marketing/website
I have now gotten a rough shell in place, however I think I won't go any further with it until we are more feature complete.

Also, I was looking at the prices for cloud storage (consumer) at both Google and Apple, and at Google it was £1.59 for 100GB, and at Apple it would be roughly the same price, (50GB is £0.99, and 200GB is £2.99). And that is insanely cheap, because if we look at something like BackBlaze and they are charging ~£6 per TB. And so I was thinking that a price of £5 for 100GB (and maybe we'd do a 15GB for £2?) a month would be fair enough, considering the other costs as well, however it's too expsensive for the average consumer? I dunno but I'm going to press ahead and hope that there is even the possibility of making back cost.