UPDATE apis/web
I changed tack with how the authorisation function worked, by getting the database to decode the json struct instead of in go. And then I got that whole function working, and now I am going to work on implementing the songs table, which shouldn't really be too hard, however before I do that I want to get rid of all the old data that is hanging about in the database, and in the docker volumes, so I am also going to wipe those, and then check/improve the first run experience / handling as well.