UPDATE apis/web
So  I have now decided that the best way forward is to create our own version of the postgres image, in this way we can bundle our init.sql with the image, rather than having to transport that same init.sql to every single installation otherwise.

And then I started testing the init process, that first time run of the service, and at first it was atrocious, what with go starting whilst the database was starting up and then process's failing because there wasn't a full database to connect to.
So to aleviate this I kagi'd it and found something called wait-for-it.sh, which allows me to wait for until postgres is ready and then start up. Perfect, I got all the niggles with that sorted out and then it was a simple case of making sure that init.sql was as it should be.
It wasn't, so I found out that by passing in the name of the database to my postgres container, that was enough to create the database, so I just removed that CREATE database statement from the top and we were golden.

And now it really is as simple as docker compose up that first time, and everything should move in concert and there should be no problems at all, meaning it's easier for me, and everybody else that should want to self host OM2.