UPDATE small things
I am going to try how it will work if i use VPS as a build server for this, because it is much quicker, and i can't be bothered waiting for all of this building nonsense, it takes so long, and don't get me started on the downloads. PLUS the build cache is taking up storage on my laptop, which i don't particularly want used up like that.