UPDATE web/apis songTitle updating
I got a little bit side tracked by trying to re-implement the signout functionality with cookies, but I returned too getting the updating of song titles working, and it well it works, and that's pretty cool. I also noticed that there was a bug where it would sort the order of songs by when they were first put into the database/updated, rather than the proper order, so now it is sorted by SongSequence, and that makes lots of sense.

It's funny too think that the really basic functionality of updating the songs title required a full database overhaul, which gave the oppurtunity to do the whole thing in go as well, however that has all now been done, and the service is much better for it.

I'm looking forward too improving other aspects of the site, and getting those killer features that were going too make this whole thing worthwhile in place.

But there are also some things which need to be added, such as storage management and such. And for wider application getting object storage options in place.