UPDATE web songs table drag experience
I have now gotten dragging of the elements to the state where visually we are 90% of the way, I still need to add some of the visual queues for it to be 100%, however the elements move about and that is all the is really truly needed. However I need to make it update in the database, and I might also have to on the frontend make the semi id's baked into the elements update as well, so that they would be in the correct order as the database expects, however it might not be neccessary, but i think it will be, because then 7-1 will be infront of 1-1, and that can't be good.

But it all looks good, and I've learned about more javascripty things, and that is always good.