UPDATE web - songs table drag
I have now gotten the dragover experience to a point where it will change the colour of the element that is being hovered over, and there was a bit of work with getting colours to be their correct form and format, however there are still a few changes needed to get the colours to go back to their original properly and fully. And this is just something that will take a little bit of work, however at the end of it drag should be really nice and a good boost for the project.

And as before I have made sure that drag events only happen if the user is a editor or above.