UPDATE web - drag
I am going to take a break with drag since there are a number of things that could use my time better and it is only a visual bug, but I still can't seem to figure out how i can set the background colour of those other items back to the default. This is something that could use a lot of time and in the future I will surely come back to it, however this is the best that can be done for now, perhaps the best course of action is to reload all of the elements, lazy but certain to be effective.