UPDATE web - upload/reload table
I have finally done something that I have been considering doing for a while, I went through and changed all instances of passing through details into functions in projectView.js and I have now replaced it with a file global of Details, and it means that I only need to update that one variable and then reload those sections of course, but then it also meant that I could tackle the different functions which were all making their own calls to the server for details, which is fixed by them now just accessing Details.

Plus I got rid of the UpdateLoadInTable function, and reworked loadInTable to be able to do the job for the both of them. This is good because now that's some code gone.

Plus I started work on getting the new upload animations / representations for the uploads in place by deleting the old stuff, which was fine for it's time, but now I am going to be going too it's more final implementation which won't be like how it is now.