UPDATE web - fonts
So for a while I have been mildly obsessed with the idea of using a different font than the ones I had before this commit, and I started looking at all of those exspensive shiny fonts with longing. However I went back to google fonts and found a better pairing which I really rather like and well it's sticking for now and I will avoid paying for a font for hopefully a long time.

And Now I want to see about getting a share/embed system working. So not the process of sharing the project with others for collaboration uses, but rather embedding the project of song in a website, and getting those opengraph tags working as well.