UPDATE iframe-share
So I have now built systems in go that allow me to manipulate a HTML string using regex to replace any all placeholders that might be needed. And that is pretty good, which means that it can now serve the iframe as needed, however there will be some work ahead of me, in allowing an iframe to be shared, the considerably largest will be figuring out how content share should work, should I duplicate the files that are being shared as I-Frames? potentially, however that is also wasteful, so instead I would potentially back link it to that iframe, but then that raises other problems, so then I could also allow the user to choose, if they would like enhanced privacy/security at the cost of storage, or not. But I think in reality a different approach is required, where we do in fact duplicate, because that is probably best all round. But then for some a pin could be required?

Anyway I will be putting some of this to the side for now, since there are more important things to be getting on with, but the groundwork is here now, and that is all that matters.