UPDATE web/apis - update pfp
I have now implemented the code, and adapted some of the other code that was left open for this type of malleability, to now allow the user to update their pfp in the settings page, I will still have to at some point create a more mobile friendly version of this settings page, however for the moment it now works, plus the upload system works using the same one that was built for the project image selector, and this is good news, because it means that the interface is consistent across these different areas for these similar things.

The next thing will be getting the songSequence order updates working on the database.