UPDATE web/apis - added new loading image
I have now added some new theme matched loading images for use in projects for when the real data hasn't arrived yet and I want the UI to load in so that there is some idea that something is happening, however I was just using the default project image which isn't so cool, so I made some new *temporary* ones.

There is however a little hold up in that the om2.is_dark() function isn't working properly at the moment, and this might be because of something that I have done wrong, however the function looks correct, so I might just wait for a week or something with this one and then see if it's fixed, but the issue is also present in function, so there might have to be some more work done to make sure it works.