UPDATE apis - prelogin console bug
Prelogin was printing when the cookie wasn't present, and this meant that the empty cookie was then passed on to authenticate, where previously it would then try and understand the cookie, however it was obviously impossible to parse an empty string. So now it will just return not authorised 401 when the cookie is not present.