UPDATE web/projectView - reduced text network updates
Previously when the user would click on, or tab between certain text fields that could be modified, it would send a network request to update those fields in the database, even if the text hadn't been updated. This update now fixes this, meaning that it will only be updated when the text has actually changed. An efficieny win.