UPDATE apis - dockerfile
I decided to explicitly state the environment variable of GIN_MODE=debug. And then I tried it also with GIN_MODE=release, and that worked, however the only thing that I still need to get in place is allowing the loading of a page from for example /projects/asd where this system wouldn't serve the index.html and such, which it really should do, but I can work on that.