UPDATE apis/web - update songs order
So I have now finally gotten the updating of the songs order on the backend fully working, and this is again with the power of SQL doing all the work for me. However I also had to find a way of updating the information on the frontend so that the UI works, however this hasn't been done in so good a way, since I tried creating a more efficient method, however that wasn't working, and at this point in time I don't feel like this minor feature should be taking up more of my time. And so with that I conclude all of these works on the song order updating.

Plus I made project deletions work.

Plus added check for if the correct file paths exist when uploading a song, this existed for images, however has now been pasted over.