UPDATE web/apis storage limits
I have now implemented a system whereby when a user attempts to upload a file, the system will check if they are exceeding their storage allowance, and if that is the case, it will not allow any uploads, and then in the javascript also it creates a notification that the user will now see, which shows that they have hit a storage limit error.

And now I just need to think of other areas that need the notification system, and it's really easy to create a notification now, so that is really rather good.