UPDATE web - login notifications
I have now changed the login animations so that instead of using those weird ones that I made a long time ago, it instead now uses the notification system that I implemented, and the next thing that I want to do with these notifications is to create a way where by I can create a special notification for showing uploads, since they would be a very handy thing to have in this same system, since uploads should very much be a standard path through, even if all three upload types when the final one arrives would of course be three seperate systems.

And I also need to add error handling for when the cookie becomes invalid, since the user needs to be informed of these things.