UPDATE web - login error handling/notifications
So I was working on login error handling, so that when there was no connection to the server the system would be able to tell the user about this. And this resulted in the notification system showing two notifications at once, and this wasn't ideal of course, so I then just used a setTimeout, which would have been goodenough, however I decided to just properly implement a buffer system, and that is what I have now, and that is really rather good, there are somethings that I would do to make the buffer better, and perhaps more universally appllicable, however for now it is good enough, and I can go on and do something else.